requirements.txt
streamlit>=1.28.0
python-binance>=1.0.17
ccxt>=4.0.0
vectorbt>=0.26.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
kaleido>=0.2.1
